cleanup code for all-in-one devstack support

+++ etc/3-report-r.sh
@@ -1,15 +1,21 @@
 #!/bin/bash
-
-set -x
-
 TOP_DIR=$(cd $(dirname "$0") && pwd)
 pushd $TOP_DIR > /dev/null
 
 result_folder="results"
+remote="usr@host"
+
+if ! [ $# -eq 1 ]
+then
+    echo "report-r.sh <time-offset>" 1>&2
+    exit 1
+fi
+
 mkdir $result_folder
 rm $result_folder/*
 cp -r openstack-bench/$result_folder/* ./$result_folder
-scp yingxin@10.239.48.9:~/openstack-bench/openstack-bench/$result_folder/* ./$result_folder
+scp $remote:~/nova-scheduler-bench/openstack-bench/$result_folder/* ./$result_folder
 python parse.py --offset $1 $result_folder
+echo "Done"
 
 popd > /dev/null
